fixed errors
formatted
