+++ shellborn/src/attribute/builder.rs
@@ -1,7 +1,9 @@
-use crate::attribute::attribute::Attribute;
+use crate::attribute::attribute_decoration::AttributeDecoration;
+use crate::attribute::Attribute;
 
 pub struct ConcatenatedAttributeBuilder {
     name: String,
+    decoration: AttributeDecoration,
     sub_attributes: Vec<Attribute>,
 }
 
@@ -9,36 +11,67 @@ impl ConcatenatedAttributeBuilder {
     pub fn new<S: Into<String>>(name: S) -> Self {
         Self {
             name: name.into(),
+            decoration: AttributeDecoration::Default,
             sub_attributes: vec![],
         }
     }
 
-    pub fn add_basic<S: Into<String>>(mut self, name: S) -> Self {
-        self.sub_attributes.push(Attribute::basic(name));
+    pub fn set_decoration(mut self, decoration: AttributeDecoration) -> Self {
+        self.decoration = decoration;
         self
     }
 
-    pub fn add_id<S: Into<String>>(mut self, name: S) -> Self {
-        self.sub_attributes.push(Attribute::id(name));
-        self
-    }
-
-    pub fn add_derived<S: Into<String>>(mut self, name: S) -> Self {
-        self.sub_attributes.push(Attribute::derived(name));
-        self
-    }
-
-    pub fn add_multi_valued<S: Into<String>>(mut self, name: S) -> Self {
-        self.sub_attributes.push(Attribute::multi_valued(name));
-        self
-    }
-
-    pub fn add_concatenated(mut self, concatenated: ConcatenatedAttributeBuilder) -> Self {
-        self.sub_attributes.push(concatenated.build());
+    pub fn add_attribute(mut self, attribute: Attribute) -> Self {
+        self.sub_attributes.push(attribute);
         self
     }
 
     pub fn build(self) -> Attribute {
-        Attribute::concatenated(self.name, self.sub_attributes)
+        Attribute {
+            name: self.name,
+            decoration: self.decoration,
+            sub_attributes: self.sub_attributes,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::attribute::attribute_decoration::AttributeDecoration;
+    use crate::attribute::builder::ConcatenatedAttributeBuilder;
+    use crate::attribute::Attribute;
+
+    #[test]
+    fn builder() {
+        assert_eq!(
+            ConcatenatedAttributeBuilder::new("concatenated")
+                .add_attribute(Attribute::basic("basic"))
+                .add_attribute(
+                    ConcatenatedAttributeBuilder::new("sub_concatenated")
+                        .add_attribute(Attribute::basic("sub_basic"))
+                        .build(),
+                )
+                .build(),
+            Attribute {
+                name: "concatenated".to_string(),
+                decoration: AttributeDecoration::Default,
+                sub_attributes: vec![
+                    Attribute {
+                        name: "basic".to_string(),
+                        decoration: AttributeDecoration::Default,
+                        sub_attributes: vec![],
+                    },
+                    Attribute {
+                        name: "sub_concatenated".to_string(),
+                        decoration: AttributeDecoration::Default,
+                        sub_attributes: vec![Attribute {
+                            name: "sub_basic".to_string(),
+                            decoration: AttributeDecoration::Default,
+                            sub_attributes: vec![],
+                        }]
+                    }
+                ]
+            }
+        );
     }
 }
