+++ shellborn/src/relation/builder.rs
@@ -1,5 +1,5 @@
 use crate::relation::connection::Connection;
-use crate::relation::relation::Relation;
+use crate::relation::Relation;
 
 pub struct RelationBuilder {
     name: String,
@@ -14,27 +14,33 @@ impl RelationBuilder {
         }
     }
 
-    pub fn add_zero_to_one<S: Into<String>>(mut self, to_name: S) -> Self {
-        self.connections.push(Connection::zero_to_one(to_name));
-        self
-    }
-
-    pub fn add_zero_to_many<S: Into<String>>(mut self, to_name: S) -> Self {
-        self.connections.push(Connection::zero_to_many(to_name));
-        self
-    }
-
-    pub fn add_one_to_one<S: Into<String>>(mut self, to_name: S) -> Self {
-        self.connections.push(Connection::one_to_one(to_name));
-        self
-    }
-
-    pub fn add_one_to_many<S: Into<String>>(mut self, to_name: S) -> Self {
-        self.connections.push(Connection::one_to_many(to_name));
+    pub fn add_connection(mut self, connection: Connection) -> Self {
+        self.connections.push(connection);
         self
     }
 
     pub fn build(self) -> Relation {
         Relation::new(self.name, self.connections)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::relation::builder::RelationBuilder;
+    use crate::relation::connection::Connection;
+    use crate::relation::Relation;
+
+    #[test]
+    fn builder() {
+        assert_eq!(
+            RelationBuilder::new("test")
+                .add_connection(Connection::one_to_many("a"))
+                .add_connection(Connection::zero_to_many("b"))
+                .build(),
+            Relation {
+                name: "test".to_string(),
+                connections: vec![Connection::one_to_many("a"), Connection::zero_to_many("b"),]
+            }
+        );
     }
 }
