+++ shellborn/src/relation/mod.rs
@@ -1,4 +1,20 @@
-pub mod relation;
-pub mod connection;
-pub mod cardinality;
 pub mod builder;
+pub mod cardinality;
+pub mod connection;
+
+use crate::relation::connection::Connection;
+
+#[derive(PartialEq, Debug)]
+pub struct Relation {
+    pub name: String,
+    pub connections: Vec<Connection>,
+}
+
+impl Relation {
+    pub fn new<S: Into<String>>(name: S, connections: Vec<Connection>) -> Self {
+        Self {
+            name: name.into(),
+            connections,
+        }
+    }
+}
